Allow packages to provide custom hints for MethodErrors
Package authors may be able to predict likely user errors, and it can
be nice to create an informative hint. For MethodErrors, we've done this
repeatedly for Base; this PR makes it possible for packages to get in
on the fun.

The procedure is described in the docstring for `Base.methoderror_hints`.