fixup-libgfortran: Bail out if libgfortran can't be found (#35096)
Previously if the library was not found, we would try to look for
the directory of an empty file, which ironicall does something almost
sensible:
```
$ gfortran -print-file-name=
/usr/lib/gcc/x86_64-linux-gnu/9/
```
Unfortunately, we then `dirname` the directory, stripping the last path
component. Often this is harmless, but on some systems, the 32bit path
is e.g. `/usr/lib/gcc/x86_64-linux-gnu/9/32/`, which can cause the
64 bit path to be copied instead of the 32 bit path. This is of course
only an issue if the appropriate runtime libraries are missing, but an
arch mismatch error is harder to debug than a missing libraries error
(as well as it being of course incorrect to copy the 64 bit libraries
to our build directory). This fixes fixup-libgfortran to no try and
look for the directory of an empty libname.