Support `⊙` and `⊗` as elementwise- and tensor-product operators
While we have broadcasting and `a*b'`, sometimes you need to pass
an operator as an argument to a function. Since we already have `dot` or `⋅`
for the inner product, these elementwise and tensor products fill out
the space of possibilities.

Note that adjoint/transpose are covectors, and so to prevent confusion
calling `u ⊗ v'` is an error.