Fix minimum/maximum over dimensions with missing values for `BigInt`
Types such as `BigInt` don't support `typemin` and `typemax` so
the current method to find an initial value different from `missing` fails.
Use the largest/smallest non-missing value to initialize the array instead.
This is an inefficient approach. Faster alternatives would be to avoid using
an initial value at all, and instead keep track of whether a value has been
set in a separate mask; or to use `typemax`/`typemin` for types that support them.