Fix backward compat for CodeInfo serialization
Here we introduce an explicit codeinfo_ver version number to track the
CodeInfo struct. Core.CodeInfo has seen multiple changes between Julia
versions and the addition of the hide_in_stacktrace flag is not possible
to detect via the types which exist in the serialized stream.

To solve this problem and to avoid backward compatibility code becoming
more convoluted and hard to maintain in the future, it seems simplest to
introduce an Int32 version flag which should be a single byte of
overhead.

An alternative which was considered was to rely on the value of the
header ser_version. However, this would need to be somehow stored with
any user-defined subtype of AbstractSerializer which would be a API
breaking change.