Update multidimensional.jl
Hello,

Locally, this fixes https://github.com/JuliaLang/julia/issues/35396 and https://github.com/JuliaLang/julia/issues/35248. I guess the recursive function definition is not inferring and the type assertion helps. It prevents the type instability to propagate further.

This is an easy/lazy fix, please feel free to close if this is not appropriate or suggest a better fix :)

Cheers!