Windows: add wrapper `.exe` to work around lack of `RPATH`
Windows has no concept of `RPATH`, which makes shipping binaries with
relatively-located dependencies quite challenging.  Despite
investigating many potential avenues of `RPATH` emulation, (including
but not limited to Application Manifests, PE file patching, and bundling
the true Julia executable as a resource inside of a launcher) the most
reliable (And least breaking for external workflows) was found to be to
create a launcher executable that invokes the true Julia executable from
within the `libexec` directory.  In order to not run afoul of PATH
length limits, we store the original path in a separate sidechannel
environment variable (`JULIA_ORIGINAL_PATH`) then restore it after
startup is complete.