Speed up `dlpath()` on MacOS (#35630)
`dlpath()` on MacOS iterated over all loaded image names, calling
`dlopen()` on them again and comparing returned handles to see if the
given handle matched.  This wasted a lot of time running through the
`dlopen()` machinery, we can bypass much of it by passing `RTLD_NOLOAD`
in to `dlopen()`, as we are assured that all pathnames we are probing
are already loaded.