Fix dropped bytes when bulk reading from uv streams with oversized buffers (#35695)
In readbytes!(s::UVStream, buf::AbstractVector{UInt8}, nb), we wrap `buf`
in a PipeBuffer of maxsize nb and swap it in as the main buffer for the
stream `s`. However, when we inform libuv of the size of the buffer,
we instead use the size of the underlying array, which can be larger.
In that case, we probably either drop bytes or override something that
the user did not want overriden.