Better codegen for unspecialized `copyto!(::Array, ::Array)`
When the eltype and/or dimensionality is unspecified, there were
2 or 3 methods that might apply here (see
`code_typed(copyto!, (Array, Array))`). Consequently partially-specified
codegen tends to be poor for this operation.

This harmonizes our implementations and eliminates eltype and dimensionality
as a consideration for the dispatch. By default, it dispatches on
`IndexStyle(src)` after a call to `unalias`, which makes sense because
copying an array can change its preferred `IndexStyle`.
It also fixes an aliasing bug.