Documentation of ambiguous syntax, operator suffix (#28525)
* Documentation of amiguous syntax, operator suffix

This updates the documentation in line with the proposal in #28441 to disallow `+ᵃx`

Ideally, the parser should be updated together with the documentation, but I'm not familiar enough with the parser internals to do that myself.

This is a breaking change, since `+ᵃx` is currently legal syntax. (The way I interpret semantic versioning, the breaking change occurs when the documentation is updated. If the parser is later updated to work in accordance with the documentation, then that would technically be a bugfix, and wouldn't have to wait until the next major release. In practice though, people will be upset if code that was working in 1.0 suddenly stops working in 1.x.)

* fix whitespace

Co-authored-by: Kristoffer Carlsson <kristoffer.carlsson@chalmers.se>