support random sampling of tuple types: add tests
EDIT: this was also implemented in #50251, which is now merged, so
merge this now for the added tests, and the added feature of
`rand(Tuple{})`.

This allows e.g `rand(NTuple{5,Int})` to sample a tuple of 5 `Int`s.

The implementation simply assembles a tuple by calling `rand` on the
corresponding type parameters of the tuple type.
A generated function is used to ensure type stability.