Split `show_ir` so line printer can be used elsewhere (#35952)
Currently JuliaInterpreter & Debugger need to print just a subset
of CodeInfo statements (those surrounding the current program counter).
Currently they print the whole thing to an IOBuffer, split the lines,
and keep just a subset. Unfortunately, because some statements
print over multiple lines, it's easy to get misaligned.

This makes it possible to call the standard printer for a specified line.