Add clearer documentation for asynchronous I/O to the manual. (#35412)
This adds a section to the end of the "networking and streams" page that
explicitly mentions asynchronous I/O and provides a couple of code
examples using the @sync / @async macros. It uses a Sockets-based
example for demonstrating the usage of the @sync / @async macros,
since it does a better job of demonstrating the asynchronous nature of
I/O functions.