Link to "noteworthy differences" early in docs
From discourse: https://discourse.julialang.org/t/the-speed-of-light-is-an-integer-why-should-we-care/40108/51?u=per

It frequently happens that newcomers to Julia are tripped up by some aspect of the language, such as integer overflow, that is not present in their current favourite programming language. The list of noteworthy differences points out many of these pitfalls, so getting people to read it can reduce user frustration.