Remove NEWS.md from the PDF of the manual (#36312)
* Partial fix for #35495

This removes the NEWS.md link at the bottom of the first page of the
manual. Given that we link to it prominently on top of the page, I
feel this is ok.

This will remove Section II of the PDF manual, where the release notes
take 15 pages (because ofthe way markdown gets converted to PDF).

* Better layout for the generated PDF

(cherry picked from commit 43fd66c8ac9010a193fbcf870ce731b99dd38b4f)