active project: get rid of the concept of a home project
This makes doing `julia --project` equivalent to activating the
project that the current directory is in at Julia startup. What
this means is that if you do `pkg> activate` instead of taking
you back to whatever project was initially active, it will make
it as if you had never activated a project in the first place.