Ensure string-hashing is defined before it gets used (#36411)
An analysis of invalidations during bootstrap revealed that `log_record_id`
calls `hash` on strings, and in principle it is used before the
specialized methods are defined. Since the new methods change the value
of the hash is expected to change, this could cause subtle bugs.
It seems safer to define the hash methods at around the first time they
could conceivably be defined.