Allow an additional corner case in GC lowering (#36540)
This allows masked loads with all one masks, which are just regular
non-masked loads, so don't need any special processing. In general,
a masked load is a load fused with a select, so it would need all
the additional handling that LiftSelect does. However, this is
sufficient for the moment and fixes #36539.