fix broadcasted assignment for scalar cartesian indexing (#35926)
* fix broadcast for scalar cartesian indexing

When broadcasting left-hand side indexing is turned into a dotview.
For scalar indexing this should be equivalent to a `getindex`, but up
until this patch the `CartesianIndex` would not be treated as such.

This fixes e.g. broadcasting like `a[CartesianIndex(1)] .= 1` for a nested
array `a`.

* Also extend to `AbstractCartesianIndex`es

Co-authored-by: Matt Bauman <mbauman@gmail.com>
(cherry picked from commit ff174120aad91e1a838b5cf99bd2b97c14aab2f5)