Rework julia loading strategy in anticipation of JLL stdlibs
We move all code from `ui/repl.c` to `libjulia` and create two new
entrypoints for the Julia REPL: an executable loader (`julia.exe`) and a
library loader (`libjulialoader.dll`).  These entrypoints know how to
load all necessary dependent libraries and set things up such that
`libjulia` can run.  `libjulialoader` is intended for use by embedding
clients.  We move this code to `cli/` as `ui/` is somewhat misleading at
this point.

All this work is in expectation of future JLL stdlib work, which will
require RPATH-like semantics on all platforms (including Windows).  This
commit enables such semantics by allowing the loader executable to
directly `dlopen()` any necessary libraries, which will include
foundational support libraries such as `libgcc_s` or `libLLVM`.  Right
now, these libraries are all dumped into `bin` or `lib` or `lib/julia`,
so finding them is not hard, however in the future they will be stored
in places such as `share/julia/stdlib/artifacts/<treehash>/lib`.

Until such a point, this PR will not force loading of anything other
than `libjulia`, `libgcc` (in order to work around FreeBSD issues) and
`libopenlibm` (in order to work around Windows issues related to
untrusted system libm).

Note that while we attempt to be as independent of system libraries as
possible, if `libpthread` is not dynamically linked to the executable
itself, when `libc` initializes its error handling routines [0] it
assumes it's running in a single-threaded mode and disables some thread
safety measures such as not using a static buffer error messsages.

Because we now do not link against libm in the executable, we must
explicitly pass `Base.libm_name` to `ccall()`.  In order to do this, we
need to re-arrange a few pieces of `base/Base.jl`, including creating a
`strcat()` from closer to first principles than we'd like.

[0] https://github.com/bminor/glibc/blob/cdf645427d176197b82f44308a5e131d69fb53ad/dlfcn/dlerror.c#L184-L188