Rm some dead code
This code is dead, but also has GC-rooting issues (it modifies
iparams with a newly allocated object, but iparams is not
guaranteed to be roots). Delete it and add an assert that
should fire if anybody ever changes the surrounding code
to make it necessary again (at which point they can find
this commit and fix it properly).