build: drop explicit support for pre-Mavericks OS in the default binaries
Hypothetically still supported in the build system back to 10.6 or so,
though I doubt it has been tested in a long time on such a system.

This lets us avoid needing to specify the libc++ explicitly, which makes
the build a bit simpler, and can help with building `clang` in our
`deps` folder (which may try to to build support libraries for foreign
targets, and get confused by these extra flags).