Override read-only permissions during `rm(path, recursive=true, force=true)`
If a directory is marked as read-only for the current user, `rm()` can
fail when a child within that directory is attempted to be removed.
This is most annoying when recursively removing directories where one of
the directories has been set to read-only.  This PR changes a recursive
remove to first set all directories to writable before removing them,
but only when `force` is set to `true`.