Fix TBAA for `DataType`
Due to lazy field type initialization, the `types` field is actually mutable in `DataType`.
This means that we cannot use immutable or constant TBAA for all `DataType` fields.

However, since any objects stored in the field are still rooted
over the lifetime of the `DataType` we can still let the GC root placement pass refine
any field to the parent `DataType`.
Also make use of the field offset to still mark known load from all other fields as constant.
This is currently done via a special case but it could also be useful for many other types too.