[mktempdir()]: reset permissions before attempting to delete
Because users may do something weird like create directories without
write permissions for the owner, it is best to do everything we can to
delete the files created within `mktempdir()`.  We add back on the write
and execute bits to all directories within the temporary path we
created, so as to be sure we can both list the directories and remove
files from the directories as we go.