Precompile `watch_file` and a few other functions used by Revise
Together with some internal changes coming in Revise v3.0, on my
machine this drops the overhead of `using Revise` from 3.5s to 0.5s.
This change only contributes about 0.1s of that time, but out of
0.5s it's a noticeable improvement. The only one of these that
makes a sizable impact is `watch_file`, but this does ensure that
Revise doesn't trigger any additional "generic" compilation.