Ensure 0 isn't padded unnecessarily with %g specifier in Printf
Fixes #37474. The issue here is that 0 is special-cased in the
`Ryu.writeshortest` algorithm, so it didn't hit the same code paths that
were already tested to ensure no unnecessary zero padding for the `%g`
specifier. We should only zero-pad when the `hash` argument is true to
match C, which this PR does.