Add `reinterpret(reshape, T, a)`
This addresses longstanding performance problems with `reinterpret`
when `sizeof(eltype(a))` is an integer multiple of `sizeof(T)`.
By reshaping the array to have an extra "channel dimension,"
LLVM can unroll the inner loop thanks to static size information.
Conversely, this consumes the initial "channel dimension" if
`sizeof(T)` is an integer multiple of `sizeof(eltype(a))`.