Fix escaped % printing in new Printf code
Fixes #37784. The classic blunder of not unescaping the escaped when
printing. I considered trying to detect escaped chars during format
construction-time, but there isn't a clean way to utilize the current
`Printf.Format` structure to achieve this, since it requires
`length(f.formats) == length(f.substringranges) + 1`, and
`f.substringranges` must be the concretely typed
`Vector{UnitRange{Int}}`. Doing the unescaping in `Printf.format` when
we're outputting seems to not incur any performance penalty and seems
like a pretty clean fix since all outputting is handled in this one
function anyway.