Omit the `N` from `SCartesianIndex2` and friends
In order to make this a subtype of `AbstractCartesianIndex`, formerly
this added a useless type parameter `N` to indicate the dimensionality
of the array for which this index was constructed. But since none of the
code depends on `N`, and it would have forced useless specialization,
it seems better to not have it.