Install GCC libraries from `CompilerSupportLibraries_jll`
Instead of bundling whatever version of GCC libraries comes with the currently
used compiler, use by default the latest version of libraries in
`CompilerSupportLibraries_jll`.  This can be disabled by setting
`USE_SYSTEM_CSL=1`.

The main benefit is that the official binaries will be able to ship a much more
recent version of GCC libraries, alleviating issues like #34276 and
https://github.com/JuliaGL/GLFW.jl/issues/198.