quite faster rand(::MersenneTwister, ::Type{Float64}) etc...
Generation of `Int64` just became faster than of `Float64` (#37914).
Given that dSFMT produces natively `Float64`, and that generating a random
`Float64` needs only 52 random bits, surely performance can be improved there.

Just ask the `gen_rand` routine, which randomizes the cache for floats,
to not inline, as it's unlikely to be called (once in a thousand).

The speedups are about 1.8x for `rand(rng, Float64)`, 13% for
`randn(rng, Float64)`, more than 1.9x for `rand(rng, Int32)`
and smaller integers (including `Bool`).